+++ velocity/src/main/java/com/atomguard/velocity/VelocityBuildInfo.java
@@ -12,7 +12,7 @@ import org.slf4j.Logger;
  */
 public final class VelocityBuildInfo {
 
-    public static final String VERSION = "2.0.6";
+    public static final String VERSION = "2.0.7";
     public static final String BUILD_DATE = "2026-03-19";
     public static final String AUTHOR = "AtomGuard Team";